responsivity adjusts
