+++ src/@types/ui/genericUiTypes.ts
@@ -1,0 +1,15 @@
+export type containerType = {
+  margin?: string;
+  font?: any;
+  padding?: string;
+  width: string;
+  height: string;
+  vAlign?: string;
+  hAlign?: string;
+  border?: string;
+  background?: string;
+  position?: string;
+  top?: string;
+  left?: string;
+  right?: string;
+};
