+++ src/ui/assets/bx-logo@2x.png
