+++ src/ui/assets/delete.png
