+++ src/ui/assets/delete@2x.png
