+++ src/ui/assets/delete@3x.png
