+++ src/ui/assets/trash.png
