+++ src/ui/assets/trash@2x.png
