+++ src/ui/assets/trash@3x.png
